+++ repos/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/KeyValueRepo.kt
@@ -94,6 +94,10 @@ suspend inline fun <Key, Value> WriteKeyValueRepo<Key, Value>.set(
 ) = set(toSet.toMap())
 
 suspend inline fun <Key, Value> WriteKeyValueRepo<Key, Value>.set(
+    toSet: List<Pair<Key, Value>>
+) = set(toSet.toMap())
+
+suspend inline fun <Key, Value> WriteKeyValueRepo<Key, Value>.set(
     k: Key, v: Value
 ) = set(k to v)
 
@@ -125,7 +129,11 @@ interface KeyValueRepo<Key, Value> : ReadKeyValueRepo<Key, Value>, WriteKeyValue
      * By default, will remove all the data of current repo using [doAllWithCurrentPaging], [keys] and [unset]
      */
     suspend fun clear() {
-        doAllWithCurrentPaging { keys(it).also { unset(it.results) } }
+        var count: Int
+        do {
+            count = count().takeIf { it < Int.MAX_VALUE } ?.toInt() ?: Int.MAX_VALUE
+            keys(FirstPagePagination(count)).also { unset(it.results) }
+        } while(count > 0)
     }
 }
 typealias StandardKeyValueRepo<Key,Value> = KeyValueRepo<Key, Value>